// src/app/models/supported-language.model.ts
export interface SupportedLanguage {
  languageCode: string;
  displayNameEN: string;  
  displayNameTH: string;
  displayNameJA: string;
  displayNameZH: string;
  displayNameIT: string;
  displayNameES: string;
  displayNameKO: string;
  displayNameID: string;
  displayNameVI: string;
  displayNameDE: string;
  displayNamePT: string;
  displayNameFR: string;
  displayNameTR: string;
  displayNameRU: string;
  nativeName: string;
  sortOrder?: number;
}